Crate kfs_kernel

KFS

A small kernel written in rust for shit and giggles. Also, hopefully the last project I'll do before graduating from 42 >_>'.

Currently doesn't do much, besides booting and printing Hello World on the screen. But hey, that's a start.

Re-exports

pub use heap_allocator::rust_oom;

Modules

__rg_allocator_abi

checks
Checked maths functions returning useful errors.

devices
Device drivers

elf_loader
Loads Kernel Built-ins.

error
UserspaceError and KernelError

event
The core event handling primitives of KFS.

frame_allocator
Physical memory manager.

heap_allocator
Virtual heap allocator.

i386
This crate is x86_64's little brother. It provides i386 specific functions and data structures, and access to various system registers.

interrupts
Interrupt handling.

ipc
IPC primitives

log_impl
A simple log implementation based on env_logger

mem
Basic functionality for dealing with memory.

paging
Paging.

process
Process

scheduler
The Completly Unfair Scheduler

sync
Synchronization primitives used by KFS

utils
Generic useful functions

Statics

ALLOCATOR

Functions

common_start
CRT0 starts here.

do_panic⚠
The kernel panic function.

eh_personality
The exception handling personality function for use in the bootstrap.

force_double_fault⚠

main

panic_fmt
Function called on panic! invocation.

start⚠
The entry point of our kernel.