Enum kfs_kernel::error::KernelError

pub enum KernelError {
    PhysicalMemoryExhaustion {
        backtrace: Backtrace,
    },
    VirtualMemoryExhaustion {
        backtrace: Backtrace,
    },
    InvalidAddress {
        address: VirtualAddress,
        length: usize,
        backtrace: Backtrace,
    },
    InvalidSize {
        size: usize,
        backtrace: Backtrace,
    },
    AlignmentError {
        given: usize,
        needed: usize,
        backtrace: Backtrace,
    },
    WouldOverflow {
        lhs: usize,
        rhs: usize,
        operation: ArithmeticOperation,
        backtrace: Backtrace,
    },
    ZeroLengthError {
        backtrace: Backtrace,
    },
    MmError(MmError),
    ProcessKilled {
        backtrace: Backtrace,
    },
    ThreadAlreadyStarted {
        backtrace: Backtrace,
    },
    InvalidCombination {
        backtrace: Backtrace,
    },
    ExceedingMaximum {
        value: u64,
        maximum: u64,
        backtrace: Backtrace,
    },
    InvalidKernelCaps {
        kcap: u32,
        backtrace: Backtrace,
    },
    IpcError {
        backtrace: Backtrace,
    },
    ReservedValue {
        backtrace: Backtrace,
    },
    ThisWillNeverHappenButPleaseDontMatchExhaustively,
}

Kernel Error.

Used pretty much everywhere that an error can occur. Holds the reason of the error, and a backtrace of its origin, for debug.

When a KernelError must be propagated to userspace, i.e. a syscall failed, it must be converted to a UserspaceError.

Variants
PhysicalMemoryExhaustion
Fields of PhysicalMemoryExhaustion

backtrace: Backtrace
VirtualMemoryExhaustion
Fields of VirtualMemoryExhaustion

backtrace: Backtrace
InvalidAddress
Fields of InvalidAddress

address: VirtualAddresslength: usizebacktrace: Backtrace
InvalidSize
Fields of InvalidSize

size: usizebacktrace: Backtrace
AlignmentError
Fields of AlignmentError

given: usizeneeded: usizebacktrace: Backtrace
WouldOverflow
Fields of WouldOverflow

lhs: usizerhs: usizeoperation: ArithmeticOperationbacktrace: Backtrace
ZeroLengthError
Fields of ZeroLengthError

backtrace: Backtrace
MmError(MmError)ProcessKilled
Fields of ProcessKilled

backtrace: Backtrace
ThreadAlreadyStarted
Fields of ThreadAlreadyStarted

backtrace: Backtrace
InvalidCombination
Fields of InvalidCombination

backtrace: Backtrace
ExceedingMaximum
Fields of ExceedingMaximum

value: u64maximum: u64backtrace: Backtrace
InvalidKernelCaps
Fields of InvalidKernelCaps

kcap: u32backtrace: Backtrace
IpcError
Fields of IpcError

backtrace: Backtrace
ReservedValue
Fields of ReservedValue

backtrace: Backtrace
ThisWillNeverHappenButPleaseDontMatchExhaustively