Struct kfs_kernel::i386::gdt::DescriptorTableEntry

#[repr(transparent)]

struct DescriptorTableEntry(u64);

An entry in the GDT/LDT.

Those entries generally describe a segment. However, the DescriptorTable also contains special descriptors called "System Descriptors". Those are used for specifying different kind of memory regions used by the CPU, such as TSS, LDT, or Call Gates.

Methods

impl DescriptorTableEntry

pub fn null_descriptor() -> DescriptorTableEntry

Returns an empty descriptor. Using this descriptor is an error and will raise a GPF. Should only be used to create a descriptor to place at index 0 of the GDT.

pub fn new(
base: u32,
limit: u32,
is_code: bool,
priv_level: PrivilegeLevel
) -> DescriptorTableEntry

Creates an empty GDT descriptor, but with some flags set correctly

pub fn new_system(
ty: SystemDescriptorTypes,
base: u32,
limit: u32,
priv_level: PrivilegeLevel
) -> DescriptorTableEntry

Creates an empty GDT system descriptor of the given type.

pub fn new_ldt(
base: &'static DescriptorTable,
priv_level: PrivilegeLevel
) -> DescriptorTableEntry

Creates a new LDT descriptor.

pub fn new_tss(
base: &'static TssStruct,
priv_level: PrivilegeLevel,
iobp_size: usize
) -> DescriptorTableEntry

Creates a GDT descriptor pointing to a TSS segment

fn get_limit(self) -> u32

Gets the byte length of the entry, minus 1.

fn set_limit(&mut self, newlimit: u32)

Sets the entry's byte length to the given number plus one. Note that if the given length is higher than 65536, it should be properly page-aligned.

Panics

Panics if the given limit is higher than 65536 and not page aligned.

fn get_base(self) -> u32

Gets the base address of the entry.

fn set_base(&mut self, newbase: u32)

Sets the base address of the entry.

pub fn get_accessed(self) -> bool

CPU sets this bit to true when the segment is accessed.

pub fn is_readwrite_allowed(self) -> bool

Code Segments: Whether read access for this segment is allowed.

Data Segments: Whether write access for this segment is allowed.

pub fn is_comformant(self) -> bool

Code Segments: if true, code in this segment can be executed from a lower privilege level (example: ring3 can far jump into ring2 code). If false, the code segment can only be executed from the right DPL.

Data Segments: if true, the segment grows up. If false, the segment grows down.

pub fn is_executable(self) -> bool

Determines whether the segment is a code segment or a data segment. If true, this is a code segment and can be executed. If false, this is a data segment.

pub fn get_ring_level(self) -> PrivilegeLevel

The privilege level associated with this segment.

pub fn get_present(self) -> bool

A segment needs to be present to have an effect. Using a not-present segment will cause an exception.

pub fn is_4k_granularity(self) -> bool

If true, the limit is a count of 4k pages. If false, it is a byte count.

fn set_4k_granularity(&mut self, is: bool)

If true, the limit is a count of 4k pages. If false, it is a byte count.

pub fn is_32bit(self) -> bool

If true, this is a 32-bit segment. If false, it is a 16-bit selector.