Module kfs_kernel::interrupts

Interrupt handling.

All exceptions are considered unrecoverable errors, and kill the process that issued it.

Feature panic-on-exception makes the kernel stop and panic when a thread generates an exception. This is useful for debugging.

Modules

irq

syscalls
Syscall implementations

Structs

IDT

Statics

IDT

Functions

alignment_check_handler

bound_range_exceeded_handler

breakpoint_handler

check_thread_killed
Checks if our thread was killed, in which case unschedule ourselves.

debug_handler

device_not_available_handler

divide_by_zero_handler

double_fault_handler

general_protection_fault_handler

init⚠
Initialize the interrupt subsystem. Sets up the PIC and the IDT.

invalid_opcode_handler

invalid_tss_handler

machine_check_handler

non_maskable_interrupt_handler

overflow_handler

page_fault_handler

panic_on_exception
Panics with an informative message.

security_exception_handler

segment_not_present_handler

simd_floating_point_handler

stack_segment_fault_handler

syscall_handler
This is the function called on int 0x80.

virtualization_handler

x87_floating_point_handler