Module kfs_kernel::interrupts::syscalls

Syscall implementations

Structs

Registers

Functions

accept_session

close_handle

connect_to_named_port

connect_to_port

create_interrupt_event

create_port

create_session

create_shared_memory

create_thread
Creates a thread in the current process. The thread can then be started with the svcStartThread.

exit_process
Kills our own process.

exit_thread
Kills our own thread.

ignore_syscall

manage_named_port

map_framebuffer
Maps the vga frame buffer mmio in userspace memory

map_shared_memory

output_debug_string

query_memory

reply_and_receive_with_user_buffer

send_sync_request_with_user_buffer

set_heap_size
Resize the heap of a process, just like a brk. It can both expand, and shrink the heap.

sleep_thread

start_thread
Starts a previously created thread.

syscall_handler_inner

unmap_shared_memory

wait_synchronization