Module kfs_kernel::ipc::session

IPC Sessions

A Session represents an established connection. It implements a rendez-vous style Remote Procedure Call interface. The ClientSession has a send_request operation, which will wait for the counterpart ServerSession's reply. A ServerSession can also receive the pending requests.

Note that a single Session can only process a single request at a time - it is an inherently sequential construct. If multiple threads attempt receiving on the same handle, they will have to wait for the current request to be replied to before being able to receive the next request in line.

use kernel::ipc::session;
let (server, client) = session::new();

The requests are encoded in a byte buffer under a specific format. For documentation on the format, switchbrew is your friend.

Structs

ClientSession
The client side of a Session.

HandleDescriptorHeader
Part of an HIPC command. Sent only when MsgPackedHdr::enable_handle_descriptor is true.

InternalSession

MsgPackedHdr
Represenens the header of an HIPC command.

Request

ServerSession
The server side of a Session.

Session

Functions

buf_map

new
Create a new Session pair. Those sessions are linked to each-other: The server will receive requests sent through the client.

pass_message