Module kfs_kernel::mem

Basic functionality for dealing with memory.

Contains definition for VirtualAddress and PhysicalAddress, and UserSpacePointer

Structs

FatPtr
Internal rust representation of a DST pointer.

PhysicalAddress
Represents a Physical address

UserSpacePtr
A pointer to read-only userspace memory. Prevents userspace from trying to use a syscall on kernel memory.

UserSpacePtrMut
A pointer to read-write userspace memory. Prevents userspace from trying to use a syscall on kernel memory.

VirtualAddress
Represents a Virtual address

Functions

count_pages
Counts the number of pages size takes

round_to_page
Rounds an address to its page address

round_to_page_upper
Rounds an address to the next page address except if its offset in that page is 0