Module kfs_kernel::paging::cross_process

Cross Process Mapping

Provides mechanics for temporarily mirroring a Userland mapping in KernelLand.

When kernel has to access memory in UserLand (for example a user has provided a buffer as an argument of a syscall), it can do it in two ways:

Either the buffer is mapped in the same page tables that the kernel is currently using, in this case it accesses it directly, through a UserspacePtr.

Either the buffer is only mapped in the page tables of another process, in which case it has to temporarily map it to KernelLand, make the modifications, and unmap it from KernelLand.

This module covers the second case.

The remapping is represented by an CrossProcessMapping structure. It is created from a reference to the mapping being mirrored, and the KernelLand address where it will be remapped. When this struct is dropped, the frames are unmap'd from KernelLand.

A CrossProcessMapping is temporary by nature, and has the same lifetime as the reference to the mapping it remaps, which is chained to the lifetime of the lock protecting ProcessMemory.

Because of this, a CrossProcessMapping cannot outlive the ProcessMemory lock guard held by the function that created it. This ensures that:

All CrossProcessMappings will be unmapped before returning to UserSpace.

Another thread cannot make any modification to a ProcessMemory while a CrossProcessMapping exists for this ProcessMemory.

The UserLand side of the mapping cannot be deleted while it is still being mirrored, as this would require a mutable borrow of the ProcessMemory lock, and it is currently (constly) borrowed by the CrossProcessMapping.

Structs

CrossProcessMapping
A struct representing a UserLand mapping temporarily mirrored in KernelSpace.