Trait kfs_kernel::paging::hierarchical_table::InactiveHierarchyTrait

pub trait InactiveHierarchyTrait: TableHierarchy {
    fn new() -> Self;

    fn switch_to(&mut self);

    fn copy_active_kernel_space(&mut self);

    fn is_currently_active(&self) -> bool;

    unsafe fn from_currently_active() -> Self;

    unsafe fn destroy(&mut self) { ... }
}

A trait implemented by innactive table hierarchies. Enables creating a

Required methods

fn new() -> Self

Creates a hierarchy. Allocates at least a top level directory, make all its entries unmapped, and make its last entry recursive.

fn switch_to(&mut self)

Switches to this hierarchy,

Since all process are supposed to have the same view of kernelspace, this function will copy the part of the active directory that is mapping kernel space tables to the directory being switched to, and then performs the switch

fn copy_active_kernel_space(&mut self)

Performs a shallow copy of the top level-directory section that maps KernelLand tables.

Used when about to switch to a hierarchy, to update it before switching to it.

fn is_currently_active(&self) -> bool

Checks if this inactive hierarchy is actually the currently active one.

Generally this means comparing the current MMU register pointer to top-level table with the address of the top-level table of this hierarchy.

unsafe fn from_currently_active() -> Self

Returns the currently active hierarchy as an inactive hierarchy.

Loading content...
Provided methods

unsafe fn destroy(&mut self)

De-allocates all physical memory used by tables of this hierarchy, by iterating in RecursiveTablesLand, and freeing every entry.

Does not unmap UserLand and KernelLand memory, this should be done before calling this function, otherwise they will be leaked.

This might be called by the Drop of the struct it's implemented on.

Loading content...