Module kfs_kernel::paging::process_memory

The management of a process' memory

j-------------------------------j j---------------------j
|        Process Memory         | |    Kernel Memory    |
j-------------------------------j j---------------------j
                |                            |
    j-----------------------j                |
    | Userspace Bookkeeping |                |
    j-----------------------j                |
                |                            |
j--------------------------------+----------------~-----j
|           User Land            |   Kernel Land  | RTL |
j--------------------------------+----------------~-----j
                        Page tables

Structs

ProcessMemory
The struct representing a process' memory, stored in the ProcessStruct behind a lock.

Enums

DynamicHierarchy

QueryMemory
Because we do not store Available mappings internally, we need this enum to return a new available mappings, or a reference to the stored mapping.