Enum kfs_kernel::process::Handle

pub enum Handle {
    ReadableEvent(Box<dyn Waitable>),
    ServerPort(ServerPort),
    ClientPort(ClientPort),
    ServerSession(ServerSession),
    ClientSession(ClientSession),
    Thread(Weak<ThreadStruct>),
    SharedMemory(Arc<Vec<PhysicalMemRegion>>),
}

A handle to a userspace-accessible resource.

Description

When the userspace manipulates a kernel construct, it does so by operating on Handles, which are analogious to File Descriptor in a Unix System. A Handle represents all kinds of kernel structures, from IPC objects to Threads and IRQ events.

A Handle may be shared across multiple processes, usually by passing it via IPC. This can be used, for instance, to share a handle to a memory region, allowing for the mapping of Shared Memory.

Most handles can be waited on via ::interrupts::syscalls::wait_synchronization, which will have relevant behavior for all the different kind of handles.

Variants
ReadableEvent(Box<dyn Waitable>)

An event on which we can wait. Could be an IRQ, or a user-generated event.

ServerPort(ServerPort)

The server side of an IPC port. See ::ipc::port for more information.

ClientPort(ClientPort)

The client side of an IPC port. See ::ipc::port for more information.

ServerSession(ServerSession)

The server side of an IPC session. See ::ipc::session for more information.

ClientSession(ClientSession)

The client side of an IPC session. See ::ipc::session for more information.

Thread(Weak<ThreadStruct>)

A thread.

SharedMemory(Arc<Vec<PhysicalMemRegion>>)

A shared memory region. The handle holds on to the underlying physical memory, which means the memory will only get freed once all handles to it are dropped.

Methods

impl Handle

pub fn as_waitable(&self) -> Result<&dyn Waitable, UserspaceError>

Gets the handle as a Waitable, or return a UserspaceError if the handle cannot be waited on.

pub fn as_client_port(&self) -> Result<ClientPort, UserspaceError>

Casts the handle as a ClientPort, or returns a UserspaceError.

pub fn as_server_session(&self) -> Result<ServerSession, UserspaceError>

Casts the handle as a ServerSession, or returns a UserspaceError.

pub fn as_client_session(&self) -> Result<ClientSession, UserspaceError>

Casts the handle as a ClientSession, or returns a UserspaceError.

pub fn as_thread_handle(&self) -> Result<Weak<ThreadStruct>, UserspaceError>

Casts the handle as a Weak<ThreadStruct>, or returns a UserspaceError.

pub fn as_shared_memory(
&self
) -> Result<Arc<Vec<PhysicalMemRegion>>, UserspaceError>

Casts the handle as an Arc<Vec<PhysicalMemRegion>, or returns a UserspaceError.