Struct kfs_kernel::process::ProcessStruct

pub struct ProcessStruct {
    pub pid: usize,
    pub name: String,
    pub pmemory: Mutex<ProcessMemory>,
    pub phandles: SpinLockIRQ<HandleTable>,
    pub threads: SpinLockIRQ<Vec<Weak<ThreadStruct>>>,
    pub killed: AtomicBool,
    pub ioports: Vec<u16>,
    thread_maternity: SpinLock<Vec<Arc<ThreadStruct>>>,
}

The struct representing a process. There's one for every process.

It contains many information about the process :

Its type (regular userspace process, or kworker)

Its state (Running, Scheduled, Stopped)

Its memory pages

Its kernel stack, for syscalls and interrupts

Its hardware context, to be restored on rescheduling

Fields
pid: usize

The unique id of this process.

name: String

A name for this process.

pmemory: Mutex<ProcessMemory>

The memory view of this process. Shared among the threads.

phandles: SpinLockIRQ<HandleTable>

The handles of this process. Shared among the threads.

threads: SpinLockIRQ<Vec<Weak<ThreadStruct>>>

The threads of this process. A ProcessStruct with no thread left will eventually be dropped.

killed: AtomicBool

Marks when the process is dying.

ioports: Vec<u16>

A vector of readable IO ports.

When task switching, the IOPB will be changed to take this into account.

thread_maternity: SpinLock<Vec<Arc<ThreadStruct>>>

An array of the created but not yet started threads.

When we create a thread, we return a handle to userspace containing a weak reference to the thread, which has not been added to the schedule queue yet. To prevent it from being dropped, we must keep at least 1 strong reference somewhere. This is the job of the maternity. It holds references to threads that no one has for now. When a thread is started, its only strong reference is removed from the maternity, and put in the scheduler, which is now in charge of keeping it alive (or not).

Note that we store in the process struct a strong reference to a thread, which itself has a strong reference to the same process struct. This creates a cycle, and we loose the behaviour of "a process is dropped when its last living thread is dropped". The non-started threads will keep the process struct alive. This makes it possible to pass a non-started thread handle to another process, and let it start it for us, even after our last living thread has died.

However, because of this, if a thread creates other threads, does not share the handles, and dies before starting them, the process struct will be kept alive indefinitely by those non-started threads that no one can start, and the process will stay that way until it is explicitly killed from outside.

Methods

impl ProcessStruct

pub fn new(name: String, ioports: Vec<u16>) -> Arc<ProcessStruct>

Creates a new process.

The created process will have no threads.

Panics

Panics if max PID has been reached, which it shouldn't have since we're the first process.

unsafe fn create_first_process() -> Arc<ProcessStruct>

Creates the very first process at boot. Called internally by create_first_thread.

The created process will have no threads.

Safety

Use only for creating the very first process. Should never be used again after that. Must be using a valid KernelStack, a valid ActivePageTables.

Panics

Panics if max PID has been reached, which it shouldn't have since we're the first process.

pub fn kill_process(this: Arc<Self>)

Kills a process by killing all of its threads.

When a thread is about to return to userspace, it checks if its state is Killed. In this case it unschedules itself instead, and its ThreadStruct is dropped.

When our last thread is dropped, our process struct is dropped with it.

We also mark the process struct as killed to prevent race condition with another thread that would want to spawn a thread after we killed all ours.