Struct kfs_kernel::process::ThreadStruct

pub struct ThreadStruct {
    pub state: ThreadStateAtomic,
    pub kstack: KernelStack,
    pub hwcontext: SpinLockIRQ<ThreadHardwareContext>,
    pub process: Arc<ProcessStruct>,
    pub int_disable_counter: AtomicUsize,
}

The struct representing a thread. A process may own multiple threads.

Fields
state: ThreadStateAtomic

The state of this thread.

kstack: KernelStack

The kernel stack it uses for handling syscalls/irqs.

hwcontext: SpinLockIRQ<ThreadHardwareContext>

The saved hardware context, for getting it running again on a process_switch.

process: Arc<ProcessStruct>

The process that this thread belongs to.

Description

A process has a link to its threads, and every thread has a link back to its process. The thread owns a strong reference to the process it belongs to, and the process in turn has a vec of weak references to the threads it owns. This way dropping a process is done automatically when its last thread is dropped.

The currently running process is indirectly kept alive by the CURRENT_THREAD global in scheduler.

int_disable_counter: AtomicUsize

Interrupt disable counter.

Description

Allows recursively disabling interrupts while keeping a sane behavior. Should only be manipulated through sync::enable_interrupts and sync::disable_interrupts.

Used by the SpinLockIRQ to implement recursive irqsave logic.

Methods

impl ThreadStruct

pub fn new(
belonging_process: &Arc<ProcessStruct>,
ep: VirtualAddress,
stack: VirtualAddress
) -> Result<Weak<Self>, KernelError>

Creates a new thread.

Sets the entrypoint and userspace stack pointer.

Adds itself to list of threads of the belonging process.

The returned thread will be in Stopped state.

The thread's only strong reference is stored in the process' maternity, and we return only a weak to it, that can directly be put in a thread_handle.

pub unsafe fn create_first_thread() -> Arc<ThreadStruct>

Creates the very first process and thread at boot.

1: Creates a process struct that uses current page tables. 2: Creates a thread struct that uses the current KernelStack as its stack, and points to the created process. 3: Adds itself to the created process.

Thread will be in state Running.

Returns the created thread.

Safety

Use only for creating the very first process. Should never be used again after that. Must be using a valid KernelStack, a valid ActivePageTables.

Does not check that the process struct is not marked killed.

Panics

Panics if max PID has been reached, which it shouldn't have since we're the first process.

pub fn start(this: Weak<Self>) -> Result<(), KernelError>

Takes a reference to a thread, removes it from the maternity, and adds it to the schedule queue.

We take only a weak reference, to permit calling this function easily with a thread_handle.

Errors

ThreadAlreadyStarted if the weak resolution failed (the thread has already been killed).

ThreadAlreadyStarted if the thread was not found in the maternity.

ProcessKilled if the process struct was tagged killed before we had time to start it.

pub fn kill(this: Arc<Self>)

Sets the thread to the Killed state.

We reschedule the thread (cancelling any waiting it was doing). In this state, the thread will die when attempting to return to userspace.

If the thread was already in the Killed state, this function is a no-op.