Module kfs_kernel::sync

Synchronization primitives used by KFS

Structs

Once
A synchronization primitive which can be used to run a one-time global initialization. Unlike its std equivalent, this is generalized so that The closure returns a value and it is stored. Once therefore acts something like 1a future, too.

RwLock
A reader-writer lock

RwLockReadGuard
A guard to which the protected data can be read

RwLockWriteGuard
A guard to which the protected data can be written

SpinLock
This type provides MUTual EXclusion based on spinning.

SpinLockGuard
A guard to which the protected data can be accessed

SpinLockIRQ
SpinLock that disables IRQ.

SpinLockIRQGuard
The SpinLockIrq lock guard.

Statics

INTERRUPT_DISARM

Traits

Lock
Abstraction around various kind of locks.

Functions

disable_interrupts
Increment the interrupt disable counter.

enable_interrupts
Decrement the interrupt disable counter.

permanently_disable_interrupts⚠
Permanently disables the interrupts. Forever.

Type Definitions

Mutex
Placeholder for future Mutex implementation.

MutexGuard
Placeholder for future Mutex implementation.