Struct kfs_kernel::sync::SpinLockIRQ

pub struct SpinLockIRQ<T: ?Sized> {
    internal: SpinLock<T>,
}

SpinLock that disables IRQ.

Description

This type behaves like a spinlock from the Linux crate. For simplicity of use and implementation. The mapping is as follows:

lock behaves like a spinlock_irqsave. It returns a guard.

Dropping the guard behaves like spinlock_irqrestore

This means that locking a spinlock disables interrupts until all spinlocks have been dropped.

Note that it is allowed to lock/unlock the locks in a different order. It uses a global counter to disable/enable interrupts. View INTERRUPT_DISABLE_COUNTER documentation for more information.

Fields
internal: SpinLock<T>
Methods

impl<T> SpinLockIRQ<T>

pub const fn new(internal: T) -> SpinLockIRQ<T>

Creates a new spinlockirq wrapping the supplied data.

pub fn into_inner(self) -> T

Consumes this SpinLockIRQ, returning the underlying data.

impl<T: ?Sized> SpinLockIRQ<T>

pub fn lock(&self) -> SpinLockIRQGuard<T>

Disables interrupts and locks the mutex.

pub fn try_lock(&self) -> Option<SpinLockIRQGuard<T>>

Disables interrupts and locks the mutex.

pub unsafe fn force_unlock(&self)

Force unlocks the lock.