Trait kfs_kernel::utils::io::Io

pub trait Io where
    <Self::Value as BitAnd<Self::Value>>::Output == Self::Value,
    <Self::Value as BitOr<Self::Value>>::Output == Self::Value,
    <Self::Value as Not>::Output == Self::Value, {
    type Value: Copy + Not + BitAnd<Self::Value> + BitOr<Self::Value> + PartialEq<Self::Value>;
    fn read(&self) -> Self::Value;

    fn write(&mut self, value: Self::Value);

    fn readf(&self, flags: Self::Value) -> bool { ... }

    fn writef(&mut self, flags: Self::Value, value: bool) { ... }
}

The Io trait allows for accessing device IO in a generic way, abstracting over different IO accesses (Port IO and Memory Mapped IO).

Associated Types

type Value: Copy + Not + BitAnd<Self::Value> + BitOr<Self::Value> + PartialEq<Self::Value>

The width of the IO access. Should be a primitive type like u8, u16, u32...

Loading content...
Required methods

fn read(&self) -> Self::Value

Reads from this Io.

fn write(&mut self, value: Self::Value)

Writes value to this Io.

Loading content...
Provided methods

fn readf(&self, flags: Self::Value) -> bool

Read from this Io, and mask the value with flags.

fn writef(&mut self, flags: Self::Value, value: bool)

Mask value with flags, and write it to this device address. Note that this causes a read!

Loading content...