Struct kfs_kernel::frame_allocator::FrameAllocator

pub struct FrameAllocator;

The physical memory manager.

Serves physical memory in atomic blocks of size PAGE_SIZE, called frames.

An allocation request returns a PhysicalMemRegion, which represents a list of physically adjacent frames. When this returned PhysicalMemRegion is eventually dropped the frames are automatically freed and can be re-served by the FrameAllocator.