Struct kfs_kernel::paging::process_memory::ProcessMemory

pub struct ProcessMemory {
    userspace_bookkeping: UserspaceBookkeeping,
    table_hierarchy: InactiveHierarchy,
    heap_base_address: VirtualAddress,
}

The struct representing a process' memory, stored in the ProcessStruct behind a lock.

We always store the table_hierarchy as an inactive hierarchy, and use a shortcut function accessing ActiveHierarchy instead if we detect it's the same cr3 as the currently active one.

A process is the only owner of a ProcessMemory

Fields
userspace_bookkeping: UserspaceBookkeeping

The list of mappings in this address space.

table_hierarchy: InactiveHierarchy

The architecture-dependent paging hierarchy.

heap_base_address: VirtualAddress

The start of the heap of this process. The heap is managed as a brk by the set_heap_size syscall.

The location of each process's heap should be random, to implement ASLR.

Methods

impl ProcessMemory

pub fn new() -> Self

Creates a ProcessMemory, allocating the userspace-bookkeeping, and the top-level table of the table hierarchy.

pub unsafe fn from_active_page_tables() -> Self

Creates a ProcessMemory referencing the current page tables. Used only when becoming the first process for creating the first ProcessMemory.

Unsafety

Having multiple ProcessMemory pointing to the same table hierarchy is unsafe.

fn get_hierarchy(&mut self) -> DynamicHierarchy

If these tables are the one currently in use, we return them as an ActiveHierarchy instead.

pub fn map_phys_region_to(
&mut self,
phys: PhysicalMemRegion,
address: VirtualAddress,
flags: MappingFlags
) -> Result<(), KernelError>

Maps a single physical regions to a given virtual address. Used to map mmio regions in UserSpace.

Error

Returns a KernelError if length is not page aligned. Returns a KernelError if length is 0.

pub fn create_regular_mapping(
&mut self,
address: VirtualAddress,
length: usize,
flags: MappingFlags
) -> Result<(), KernelError>

Allocates the physical regions, and maps them to specified address.

Error

Returns a KernelError if there was already a mapping in the range. Returns a KernelError if address does not fall in UserLand. Returns a KernelError if address or length is not PAGE_SIZE aligned.

pub fn map_shared_mapping(
&mut self,
shared_mapping: Arc<Vec<PhysicalMemRegion>>,
address: VirtualAddress,
flags: MappingFlags
) -> Result<(), KernelError>

Maps a previously created shared mapping to specified address.

Error

Returns a KernelError if there was already a mapping in the range. Returns a KernelError if address does not fall in UserLand. Returns a KernelError if address or length is not PAGE_SIZE aligned.

pub fn guard(
&mut self,
address: VirtualAddress,
length: usize
) -> Result<(), KernelError>

Guards a range of addresses

Error

Returns a KernelError if there was already a mapping in the range. Returns a KernelError if address does not fall in UserLand. Returns a KernelError if address or length is not PAGE_SIZE aligned.

pub fn unmap(
&mut self,
address: VirtualAddress,
length: usize
) -> Result<Mapping, KernelError>

Deletes a mapping in the page tables.

This function will never split an existing mapping, thus address and length must match exactly.

If the range maps physical memory, it will be de-allocated.

Error

Returns a KernelError if there was no mapping corresponding to the range. Returns a KernelError if address does not fall in UserLand. Returns a KernelError if address or length is not PAGE_SIZE aligned.

pub fn query_memory(
&self,
address: VirtualAddress
) -> Result<QueryMemory, KernelError>

Reads the state of the mapping at a given address

Error

Returns a KernelError if address does not fall in UserLand.

pub fn shrink_mapping(
&mut self,
address: VirtualAddress,
new_size: usize
) -> Result<Option<Mapping>, KernelError>

Shrink the mapping at address to new_size.

If new_size == 0, the mapping is unmapped entirely.

The removed part is returned, it is no longer mapped.

If new_size is equal to old size, nothing is done, and Ok(None) is returned.

Because it is reference counted, a Shared mapping cannot be resized.

Error

WasAvailable if address does not match any existent mapping.

InvalidSize if new_size > previous mapping length.

InvalidSize if new_size is not PAGE_SIZE aligned.

SharedMapping if address falls in a shared mapping.

InvalidMapping if address falls in a system reserved mapping.

pub fn expand_mapping(
&mut self,
address: VirtualAddress,
new_size: usize
) -> Result<(), KernelError>

Expand the mapping at address to new_size.

If the mapping used to map physical memory, new frames are allocated to match new_size.

If new_size is equal to old size, nothing is done.

Because it is reference counted, a Shared mapping cannot be resized.

Error

WasAvailable if address does not match any existent mapping.

SharedMapping if address falls in a shared mapping.

InvalidMapping if address falls in a system reserved mapping.

InvalidSize if new_size < previous mapping length.

InvalidSize if new_size is not PAGE_SIZE aligned.

InvalidSize if [address..new_size] does not fall in UserLand.

WasOccupied if a mapping was already present in the expanding area.

pub fn find_available_space(
&self,
length: usize
) -> Result<VirtualAddress, KernelError>

Finds a hole in virtual space at least length long.

Error

Returns a KernelError if no sufficiently big hole was found. Returns a KernelError if length is 0.

pub fn mirror_mapping(
&self,
address: VirtualAddress,
length: usize
) -> Result<CrossProcessMapping, KernelError>

Retrieves the mapping that address falls into, and mirror it in KernelLand

Error

Returns an Error if the mapping is Available/Guarded/SystemReserved, as there would be no point to remap it, and dereferencing the pointer would cause the kernel to page-fault.

pub fn resize_heap(
&mut self,
new_size: usize
) -> Result<VirtualAddress, KernelError>

Resize the heap of this process, just like a brk. It can both expand or shrink the heap.

If new_size == 0, it is completely de-allocated.

Return

The address of the start of the heap.

Error

InvalidSize if new_size is not PAGE_SIZE aligned.

InvalidSize if [address..new_size] does not fall in UserLand, or overlaps an existing mapping.

pub fn switch_to(&mut self)

Switches to this process memory